+++ share_server/locker-dao/src/main/java/com/share/locker/bo/OperationConfigBO.java
@@ -7,30 +7,24 @@ package com.share.locker.bo;
  *
  */
 public class OperationConfigBO extends BaseBO {
-	private Integer id;
+	private Long id;
 	private String configCode;
 	private String content;
-
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
-
 	public String getConfigCode() {
 		return configCode;
 	}
-
 	public void setConfigCode(String configCode) {
 		this.configCode = configCode;
 	}
-
 	public String getContent() {
 		return content;
 	}
-
 	public void setContent(String content) {
 		this.content = content;
 	}
