+++ share_server/locker-dao/src/main/java/com/share/locker/bo/OrderBO.java
@@ -9,104 +9,83 @@ import java.util.Date;
  *
  */
 public class OrderBO extends BaseBO {
-	private Integer orderId;
-	private Integer itemId;
-	private Integer ownerId;
-	private Integer renterId;
+	private Long orderId;
+	private Long itemId;
+	private Long ownerId;
+	private Long renterId;
 	private Integer deposit;
 	private String title;
 	private Date takeTime;
 	private Integer fee;
-	private Integer returnLockerId;
+	private Long returnLockerId;
 	private Date returnTime;
 	private Date checkTime;
-
-	public Integer getOrderId() {
+	public Long getOrderId() {
 		return orderId;
 	}
-
-	public void setOrderId(Integer orderId) {
+	public void setOrderId(Long orderId) {
 		this.orderId = orderId;
 	}
-
-	public Integer getItemId() {
+	public Long getItemId() {
 		return itemId;
 	}
-
-	public void setItemId(Integer itemId) {
+	public void setItemId(Long itemId) {
 		this.itemId = itemId;
 	}
-
-	public Integer getOwnerId() {
+	public Long getOwnerId() {
 		return ownerId;
 	}
-
-	public void setOwnerId(Integer ownerId) {
+	public void setOwnerId(Long ownerId) {
 		this.ownerId = ownerId;
 	}
-
-	public Integer getRenterId() {
+	public Long getRenterId() {
 		return renterId;
 	}
-
-	public void setRenterId(Integer renterId) {
+	public void setRenterId(Long renterId) {
 		this.renterId = renterId;
 	}
-
 	public Integer getDeposit() {
 		return deposit;
 	}
-
 	public void setDeposit(Integer deposit) {
 		this.deposit = deposit;
 	}
-
 	public String getTitle() {
 		return title;
 	}
-
 	public void setTitle(String title) {
 		this.title = title;
 	}
-
 	public Date getTakeTime() {
 		return takeTime;
 	}
-
 	public void setTakeTime(Date takeTime) {
 		this.takeTime = takeTime;
 	}
-
 	public Integer getFee() {
 		return fee;
 	}
-
 	public void setFee(Integer fee) {
 		this.fee = fee;
 	}
-
-	public Integer getReturnLockerId() {
+	public Long getReturnLockerId() {
 		return returnLockerId;
 	}
-
-	public void setReturnLockerId(Integer returnLockerId) {
+	public void setReturnLockerId(Long returnLockerId) {
 		this.returnLockerId = returnLockerId;
 	}
-
 	public Date getReturnTime() {
 		return returnTime;
 	}
-
 	public void setReturnTime(Date returnTime) {
 		this.returnTime = returnTime;
 	}
-
 	public Date getCheckTime() {
 		return checkTime;
 	}
-
 	public void setCheckTime(Date checkTime) {
 		this.checkTime = checkTime;
 	}
 
+	
 }
