+++ share_server/locker-dao/src/main/java/com/share/locker/bo/SettingBO.java
@@ -7,39 +7,31 @@ package com.share.locker.bo;
  *
  */
 public class SettingBO extends BaseBO {
-	private Integer id;
+	private Long id;
 	private String type;
 	private String key;
 	private String value;
-
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
-
 	public String getType() {
 		return type;
 	}
-
 	public void setType(String type) {
 		this.type = type;
 	}
-
 	public String getKey() {
 		return key;
 	}
-
 	public void setKey(String key) {
 		this.key = key;
 	}
-
 	public String getValue() {
 		return value;
 	}
-
 	public void setValue(String value) {
 		this.value = value;
 	}
